Added a splash screen & GUI launch menu
- This makes it _seem_ like I didn't forget that I initially promised a
  GUI interface for the OS
- Not all the programs can be run from the GUI but the GUI can execute
  SeeSH
- I modified the gl_enter()/gl_exit() commands to prevent GUI
  switch-over delays when not necessary
- Added a cool splash screen because I'm cool

+++ src/usr/seesh.c
@@ -3,15 +3,19 @@
 **
 ** Author:  Schuyler Martin <sam8050@rit.edu>
 **
-** Description: SeeSH shell program for SeeGOL
+** Description: SeeSH shell program for SeeGOL. This has since been expanded to
+**              include 
 */
 
 /** Headers    **/
 #include "../kern/gcc16.h"
 #include "seesh.h"
 
-#include "../kern/debug.h"
+#include "../kern/clock.h"
 #include "../kern/kio.h"
+#include "../gl/gl_lib.h"
+#include "../gl/img_fids.h"
+#include "../gl/pane.h"
 // definitions for all programs
 #include "program.h"
 
@@ -23,9 +27,11 @@
 #include "usr_clock.h"
 
 /** Macros    **/
-// number of shell commands. This is kept in the C file out of convenience
-#define BUILTIN_COUNT   4
+// number of programs, in various categories
+#define BUILTIN_COUNT   5
 #define PROG_COUNT      (BUILTIN_COUNT + 5)
+// number of programs that can be launched from the GUI
+#define GUI_PROG_COUNT  6
 // macros for installing programs to SeeSH
 #define INSTALL_BUILTIN_PROG(n, d, u, m) \
     prog_lst->name = n;\
@@ -37,6 +43,10 @@
 
 // structure that holds all program information
 static Program prog_lst[PROG_COUNT];
+// partial program listing used for gui menu
+static Program* gui_prog_lst;
+// display names for GUI programs in the GUI menu
+char* gui_prog_disp[GUI_PROG_COUNT];
 
 /************************** Built-in Main Functions **************************/
 
@@ -89,6 +99,15 @@ static uint8_t _clear_main(uint8_t argc, char* argv[])
     return EXIT_SUCCESS;
 }
 
+/*
+** Wrapper function that allows the GUI or SeeSH to launch itself
+*/
+static uint8_t _seesh_main(uint8_t argc, char* argv[])
+{
+    return seesh_main();
+}
+
+
 /************************** Local Helper Functions **************************/
 
 /*
@@ -123,33 +142,49 @@ static uint8_t __parse_args(char* buff, char* argv[])
 }
 
 /*
-** Initialize the shell with program structure information
+** Initialize the shell with program structure information. Attempts some
+** protection from adding to the list, which will overflow the buffer
 **
-** @param prog_lst Array of program structures
+** @param prog_lst Array of program structs; passed-in to protect 1st address
 */
-static void __init(Program* prog_lst)
+static void __seesh_init(Program* prog_lst)
 {
-    // built-in commands
-    INSTALL_BUILTIN_PROG("exit", "Bail from SeeSH", "", NULL);
-    INSTALL_BUILTIN_PROG("reboot", "Reboot the system.", "", NULL);
-    INSTALL_BUILTIN_PROG("clear", "Clears the screen.", "", &_clear_main);
-    INSTALL_BUILTIN_PROG("help", "Help menu. Describes other programs.",
-        "[program]", &_help_main);
-    // user programs
-    INSTALL_USR_PROG(hsc_tp_init);
-    INSTALL_USR_PROG(slidedeck_init);
-    INSTALL_USR_PROG(slideshow_init);
-    INSTALL_USR_PROG(trench_run_init);
-    INSTALL_USR_PROG(usr_clock_init);
+    // cheap way to prevent re-initialization of the list
+    if (!kio_strcmp(prog_lst[0].name, "exit"))
+    {
+        // standard built-in commands
+        INSTALL_BUILTIN_PROG("exit", "Bail from SeeSH", "", NULL);
+        INSTALL_BUILTIN_PROG("clear", "Clears the screen.", "", &_clear_main);
+        INSTALL_BUILTIN_PROG("help", "Help menu. Describes other programs.",
+            "[program]", &_help_main);
+        // user programs
+        INSTALL_USR_PROG(hsc_tp_init);
+        /** programs below this line will be used in the GUI display menu **/
+        gui_prog_lst = prog_lst;
+        INSTALL_USR_PROG(slidedeck_init);
+        INSTALL_USR_PROG(slideshow_init);
+        INSTALL_USR_PROG(trench_run_init);
+        INSTALL_USR_PROG(usr_clock_init);
+        // built-ins that can also be called from the GUI
+        INSTALL_BUILTIN_PROG("seesh", "SeeGOL's Shell", "", &_seesh_main);
+        INSTALL_BUILTIN_PROG("reboot", "Reboot the system.", "", NULL);
+        // namesoptions are a partial list of programs; running the shell will
+        // actually deal with this
+        for(uint8_t i=0; i<GUI_PROG_COUNT; ++i)
+            gui_prog_disp[i] = gui_prog_lst[i].name;
+    }
 }
 
 /*
 ** Main execution point of the shell
+**
+** @return exit code
 */
 uint8_t seesh_main(void)
 {
+    kio_clrscr();
     kio_print(MSG_SHELL_START);
-    __init(prog_lst);
+    __seesh_init(prog_lst);
     // exit code for the shell
     uint8_t seesh_code = EXIT_SUCCESS;
     // shell control loop
@@ -164,7 +199,7 @@ uint8_t seesh_main(void)
         // reboot and shutdown are handled by the kernel's main
         if (kio_strcmp(prog_lst[0].name, prompt_buff))
             break;
-        else if (kio_strcmp(prog_lst[1].name, prompt_buff))
+        else if (kio_strcmp(prog_lst[PROG_COUNT - 1].name, prompt_buff))
         {
             seesh_code = SYSTEM_REBOOT;
             break;
@@ -183,6 +218,9 @@ uint8_t seesh_main(void)
             if (kio_strcmp(prog_lst[id].name, argv[0]))
             {
                 err_code = prog_lst[id].main(argc, argv);
+                // pass the buck from previous SeeSH calls
+                if ((id == (PROG_COUNT - 2)) && (err_code == SYSTEM_REBOOT))
+                    return err_code;
                 break;
             }
         }
@@ -209,4 +247,71 @@ uint8_t seesh_main(void)
     }
     kio_print("SeeSH: Good-bye!\n");
     return seesh_code;
+}
+
+/***** GUI Shell *****/
+
+/*
+** System boot splash screen, like the good ol' days
+*/
+void seesh_splash(void)
+{
+    // default to VGA13
+    gl_enter(VGA_MODE_13);
+
+    gl_draw_rect_wh(PT2(0, 0), gl_getw(), gl_geth(), RGB_WHITE);
+    gl_draw_img_center_scale(IMG_FID_HSC, 3);
+    Point_2D str_ul = {0, 0};
+    Point_2D str_bb;
+    gl_draw_str_bb(str_ul, "SeeGOL", 2, gl_getw(), &str_bb);
+    str_ul.x = (gl_getw() - str_bb.x) / 2;
+    str_ul.y = gl_geth() - (str_bb.y + (str_bb.y / 2));
+    gl_draw_str_scale(str_ul, RGB_HSC, RGB_HSC, "SeeGOL", 2, gl_getw());
+    // "It really gets [the crowd excited] when they see this ramp just slowly 
+    // extending down." - Rick Sanchez
+    clk_busy_wait(4);
+    gl_clrscr();
+}
+
+/*
+** Main execution point of the "visual shell". This has limited access to
+** programs on SeeGOL. SeeSH can be launched from this GUI.
+**
+** @return exit code
+*/
+uint8_t seesh_gui_main(void)
+{
+    __seesh_init(prog_lst);
+
+    while(true)
+    {
+        // default to VGA13
+        pane_enter(VGA_MODE_13);
+        // user picks from the list of programs
+        uint8_t pid = pane_draw_prompt("SeeGOL Programs", GUI_PROG_COUNT,
+            gui_prog_disp);
+
+        // check for reboot command
+        if (pid == (GUI_PROG_COUNT - 1))
+        {
+            pane_exit();
+            return SYSTEM_REBOOT;
+        }
+        // we have to be careful with SeeSH and tear down GUI
+        else if (pid == (GUI_PROG_COUNT - 2))
+        {
+            pane_exit();
+            // pass the buck
+            if (gui_prog_lst[pid].main(1, NULL) == SYSTEM_REBOOT)
+                return SYSTEM_REBOOT;
+        }
+        // programs that can be launched from the GUI will have to handle the
+        // lack of command line args
+        else
+            gui_prog_lst[pid].main(1, NULL);
+    }
+
+    // leave graphics mode
+    pane_exit();
+    return EXIT_SUCCESS;
 }
